+++ core/io/Ui_output_option.py
@@ -2,7 +2,7 @@
 
 # Form implementation generated from reading ui file 'core/io/output_option.ui'
 #
-# Created by: PyQt5 UI code generator 5.12
+# Created by: PyQt5 UI code generator 5.12.2
 #
 # WARNING! All changes made in this file will be lost!